Refactored how settings are handled for easier testing
