+++ mcts/src/test/java/org/motechproject/nms/mcts/handler/MctsImportJobHandlerTest.java
@@ -1,11 +1,13 @@
 package org.motechproject.nms.mcts.handler;
 
+import org.joda.time.LocalDate;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import org.motechproject.commons.date.util.DateUtil;
 import org.motechproject.event.MotechEvent;
 import org.motechproject.nms.mcts.exception.MctsImportConfigurationException;
 import org.motechproject.nms.mcts.service.MctsWsImportService;
@@ -13,9 +15,14 @@ import org.motechproject.nms.mcts.utils.Constants;
 import org.motechproject.scheduler.contract.CronSchedulableJob;
 import org.motechproject.scheduler.service.MotechSchedulerService;
 import org.motechproject.server.config.SettingsFacade;
+import org.motechproject.testing.utils.TimeFaker;
 
+import java.net.URL;
+
+import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
@@ -69,7 +76,21 @@ public class MctsImportJobHandlerTest {
 
     @Test
     public void shouldStartImport() {
-        mctsImportJobHandler.handleImportEvent(new MotechEvent());
-        verify(mctsWsImportService).importFromMcts();
+        final LocalDate today = DateUtil.today();
+        final LocalDate yesterday = today.minusDays(1);
+        try {
+            TimeFaker.fakeToday(today);
+
+            when(settingsFacade.getProperty(Constants.MCTS_LOCATIONS)).thenReturn("4,15,51,2");
+            when(settingsFacade.getProperty(Constants.MCTS_ENDPOINT)).thenReturn("http://localhost:9090/test.svc");
+
+            mctsImportJobHandler.handleImportEvent(new MotechEvent());
+
+            ArgumentCaptor<URL> urlCaptor = ArgumentCaptor.forClass(URL.class);
+            verify(mctsWsImportService).importFromMcts(eq(asList(4L, 15L, 51L, 2L)), eq(yesterday), urlCaptor.capture());
+            assertEquals("http://localhost:9090/test.svc", urlCaptor.getValue().toString());
+        } finally {
+            TimeFaker.stopFakingTime();
+        }
     }
 }
